route cleanup & security
