+++ hamza-server/src/api/custom/get-verification-status/route.ts
@@ -4,12 +4,20 @@ import { RouteHandler } from '../../route-handler';
 
 export const GET = async (req: MedusaRequest, res: MedusaResponse) => {
     const handler: RouteHandler = new RouteHandler(
-        req, res, 'GET', '/custom/get-verification-status'
+        req, res, 'GET', '/custom/get-verification-status', ['customer_id']
     );
 
     await handler.handle(async () => {
+        //validate 
+        if (!handler.requireParams(['customer_id']))
+            return;
+
+        //security 
+        if (!handler.enforceCustomerId(handler.inputParams.customer_id))
+            return;
+
         let customerData = await CustomerRepository.findOne({
-            where: { id: req.query.customer_id.toString() },
+            where: { id: handler.inputParams.customer_id.toString() },
             select: { is_verified: true },
         });
