+++ hamza-server/src/api/custom/notification/add/route.ts
@@ -1,6 +1,5 @@
 import type { MedusaRequest, MedusaResponse, Logger } from '@medusajs/medusa';
 import CustomerNotificationService from '../../../../services/customer-notification';
-import { readRequestBody } from '../../../../utils/request-body';
 import { RouteHandler } from '../../../route-handler';
 
 export const POST = async (req: MedusaRequest, res: MedusaResponse) => {
@@ -16,25 +15,19 @@ export const POST = async (req: MedusaRequest, res: MedusaResponse) => {
     );
 
     await handler.handle(async () => {
+        //validate 
+        if (!handler.requireParams(['customer_id']))
+            return;
+
+        //security 
+        if (!handler.enforceCustomerId(handler.inputParams.customer_id))
+            return;
+
         const types = await customerNotificationService.addOrUpdateNotification(
             handler.inputParams.customer_id,
             handler.inputParams.notification_type
         );
 
-        res.status(200).json({ types });
+        res.status(201).json({ types });
     });
-
-    // try {
-    //     const types = await customerNotificationService.addOrUpdateNotification(
-    //         customer_id,
-    //         notification_type
-    //     );
-
-    //     res.status(200).json({ types });
-    // } catch (err) {
-    //     logger.error('Error creating notification types', err);
-    //     res.status(500).json({
-    //         error: 'Failed to create notification types',
-    //     });
-    // }
 };
