+++ hamza-server/src/api/custom/store/route.ts
@@ -1,15 +1,21 @@
 import type { MedusaRequest, MedusaResponse, Logger } from '@medusajs/medusa';
-import { readRequestBody } from '../../../utils/request-body';
 import { RouteHandler } from '../../route-handler';
+import ProductService from 'src/services/product';
 
 export const GET = async (req: MedusaRequest, res: MedusaResponse) => {
-    const customerService = req.scope.resolve('customerService');
+    const productService: ProductService = req.scope.resolve('productService');
 
-    const handler: RouteHandler = new RouteHandler(req, res, 'GET', '/custom/store');
+    const handler: RouteHandler = new RouteHandler(
+        req, res, 'GET', '/custom/store', ['product_id']
+    );
 
     await handler.handle(async () => {
-        const customers = await customerService.findAllCustomers();
-        return res.json({ customers });
+        //validate 
+        if (!handler.requireParams(['product_id']))
+            return;
+
+        const store_name = await productService.getStoreFromProduct(handler.inputParams.product_id);
+        res.json(store_name);
     });
 };
